+++ src/PockyBot.NET/Constants/Commands.cs
@@ -11,5 +11,7 @@ namespace PockyBot.NET.Constants
         public const string Reset = "reset";
         public const string Rotation = "rotation";
         public const string LocationConfig = "locationconfig";
+        public const string StringConfig = "stringconfig";
+        public const string RemoveUser = "removeuser";
     }
 }