+++ src/PockyBot.NET/Persistence/Repositories/IConfigRepository.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using System.Threading.Tasks;
 using PockyBot.NET.Persistence.Models;
 
 namespace PockyBot.NET.Persistence.Repositories
@@ -9,5 +10,7 @@ namespace PockyBot.NET.Persistence.Repositories
         int? GetGeneralConfig(string name);
         IList<StringConfig> GetAllStringConfig();
         IList<string> GetStringConfig(string name);
+        Task AddStringConfig(string name, string value);
+        Task DeleteStringConfig(string name, string value);
     }
 }