+++ src/PockyBot.NET/StartupExtensions.cs
@@ -44,6 +44,8 @@ namespace PockyBot.NET
             services.AddTransient<ITrigger, Reset>();
             services.AddTransient<ITrigger, Rotation>();
             services.AddTransient<ITrigger, LocationConfig>();
+            services.AddTransient<ITrigger, StringConfig>();
+            services.AddTransient<ITrigger, RemoveUser>();
             services.AddTransient<ITrigger, Default>();
         }
     }